requirement2.txt
qdrant_client==1.9.0
Requests==2.31.0
semantic_text_splitter==0.12.3
strenum==0.4.15
trafilatura==1.8.1
tokenizers==0.19.1
llama-index-readers-file
llama-index-readers-web==0.1.9
llama-index-vector-stores-upstash
llama-index-vector-stores-qdrant
#llama-index-vector-stores-elasticsearch
#llama-index-llms-anthropic
#llama-index-llms-cohere
#llama-index-llms-gradient
llama-index-llms-groq
#llama-index-llms-perplexity
#llama-index-llms-openai
llama-index-llms-gemini==0.1.7
llama-index-llms-together==0.1.3
llama-index-llms-ollama
#llama-index-llms-llama-api
google-generativeai
llama-index-embeddings-together
#llama-index-embeddings-voyageai
#llama-index-embeddings-openai
llama-index-embeddings-cohere
llama-index-vector-stores-chroma  